add pure autowire
