+++ macros/src/main/scala-2/com/softwaremill/macwire/MacwireMacros.scala
@@ -149,4 +149,96 @@ object MacwireMacros {
     }
   }
 
+  def autowire_impl[T: c.WeakTypeTag](
+      c: blackbox.Context
+  )(dependencies: c.Expr[Any]*): c.Expr[T] = {
+    import c.universe._
+
+    val targetType = implicitly[c.WeakTypeTag[T]]
+    lazy val typeCheckUtil = new TypeCheckUtil[c.type](c, log)
+
+    sealed trait Provider {
+      def `type`: Type
+    }
+
+    case class Instance(value: Tree) extends Provider {
+      lazy val `type`: Type = typeCheckUtil.typeCheckIfNeeded(value)
+      lazy val ident: Tree = value
+    }
+
+    class FactoryMethod(params: List[ValDef], fun: Tree) extends Provider {
+
+      def result(resolver: Type => Tree): Instance = new Instance(applyWith(resolver))
+    
+      lazy val `type`: Type = fun.symbol.asMethod.returnType
+
+      def applyWith(resolver: Type => Tree): Tree = {
+        val values = params.map { case ValDef(_, name, tpt, rhs) =>
+          resolver(typeCheckUtil.typeCheckIfNeeded(tpt))
+        }
+
+        q"$fun(..$values)"
+      }
+
+    }
+
+    object FactoryMethod {
+      def fromTree(tree: Tree): Option[FactoryMethod] = tree match {
+        // Function with two parameter lists (implicit parameters) (<2.13)
+        case Block(Nil, Function(p, Apply(Apply(f, _), _))) => Some(new FactoryMethod(p, f))
+        case Block(Nil, Function(p, Apply(f, _)))           => Some(new FactoryMethod(p, f))
+        // Function with two parameter lists (implicit parameters) (>=2.13)
+        case Function(p, Apply(Apply(f, _), _)) => Some(new FactoryMethod(p, f))
+        case Function(p, Apply(f, _))           => Some(new FactoryMethod(p, f))
+        // Other types not supported
+        case _ => None
+      }
+
+    }
+
+    def providerFromExpr(expr: Expr[Any]): Provider = {
+      val tree = expr.tree
+      FactoryMethod.fromTree(tree)
+        .getOrElse(new Instance(tree))
+    }
+
+    val providers = dependencies.map(providerFromExpr)
+
+    log(s"PURE exprs: s[${dependencies.mkString(", ")}]")
+    log(s"PURE Providers: [${providers.mkString(", ")}]")
+    log(s"PURE ProvidersTYPES: [${providers.map(_.`type`).mkString(", ")}]")
+
+    def findProvider(tpe: Type): Option[Tree] = providers.find(_.`type` <:< tpe).map {
+      case i: Instance => i.ident
+      case fm: FactoryMethod => fm.result(findProvider(_).getOrElse(c.abort(c.enclosingPosition, "TODO3"))).ident
+    }
+
+    def isWireable(tpe: Type): Boolean = {
+      val name = tpe.typeSymbol.fullName
+
+      !name.startsWith("java.lang.") && !name.startsWith("scala.")
+    }
+
+    lazy val resolutionWithFallback: (Symbol, Type) => Tree = (_, tpe) =>
+      if (isWireable(tpe)) findProvider(tpe).getOrElse(go(tpe))
+      else c.abort(c.enclosingPosition, s"Cannot find a value of type: [${tpe}]")
+
+    def go(t: Type): Tree = {
+
+      val r =
+        (ConstructorCrimper.constructorTree(c, log)(t, resolutionWithFallback) orElse CompanionCrimper
+          .applyTree(c, log)(t, resolutionWithFallback)) getOrElse
+          c.abort(c.enclosingPosition, s"Failed for [$t]")
+
+      log(s"Constructed [$r]")
+      r
+    }
+
+    
+    val code = go(targetType.tpe)
+    log(s"Code: [$code]")
+
+    c.Expr[T](code)
+  }
+
 }
